refactor: remove types from components
