+++ src/components/AboutPage/about-page.tsx
@@ -1,6 +1,7 @@
+import type { FC } from 'react';
 import './about-page.css';
 
-export const AboutPage = () => {
+export const AboutPage: FC = () => {
   return (
     <div className="about">
       <h2 className="about-title">
