+++ src/components/error-boundary/error-boundary.types.ts
@@ -1,0 +1,10 @@
+import type { ReactNode } from 'react';
+
+export interface ErrorBoundaryProps {
+  fallback: ReactNode;
+  children: ReactNode;
+}
+
+export interface ErrorBoundaryState {
+  hasError: boolean;
+}
