+++ src/components/main/card-list/card/card.types.ts
@@ -1,0 +1,13 @@
+import type { Ability, Pokemon, Stat } from '../../../../interfaces/interfaces';
+
+export interface CardProps {
+  pokemon: Pokemon;
+  currentPage: number;
+}
+
+export interface IFavoriteCard {
+  name: string;
+  baseExp: number;
+  stats: Stat[];
+  abilities: Ability[];
+}
