+++ src/components/main/card-list/pagination-controls/pagination-controls.types.ts
@@ -1,0 +1,9 @@
+type paginationDirection = 'prev' | 'next';
+
+export interface IPaginationControlsProps {
+  handler: (drection: paginationDirection) => void;
+  disabled: {
+    prev: boolean;
+    next: boolean;
+  };
+}
