+++ src/components/skeleton/skeleton.types.ts
@@ -1,0 +1,6 @@
+export type SkeletonProps = {
+  count: number;
+  width?: string;
+  height?: string;
+  margin?: string;
+};
